Remove the notify worker and use errGroup
The following switches from a custom worker implementation and instead
use the sync/errGroup type. This does a nice job at ensuring that we
fan out and fan back in with the right setup.
